soar/requirements.txt
pyyaml>=6.0
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
jinja2>=3.1.0
slack-sdk>=3.26.0
jira>=3.6.0
